rename and swipe up package
