+++ api-project/package.json
@@ -8,41 +8,30 @@
     "test": "./__tests"
   },
   "scripts": {
-    "test": "mocha -r ts-node/register './__tests/*.ts' --timeout 15000"
+    "testLocal": "mocha -r ts-node/register './__tests/*.ts' --timeout 15000",
+    "testCloud": "mocha -r ts-node/register './__tests/*.ts' --reporter mochawesome --timeout 15000 --reporter-options reportFilename=report"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/MahePrime/kibies.git",
     "bugsUrl": "https://github.com/MahePrime/kibies/issues"
   },
-  "engines": {
-    "node": "18"
-  },
   "devDependencies": {
-    "@commitlint/cli": "^17.4.2",
-    "@commitlint/config-conventional": "^17.4.2",
     "@types/chai": "^4.3.4",
     "@types/chai-json-schema": "^1.4.6",
     "@types/mocha": "^10.0.1",
     "@types/node": "^18.11.18",
     "@types/supertest": "^2.0.12",
-    "@typescript-eslint/eslint-plugin": "^5.48.2",
-    "@typescript-eslint/parser": "^5.48.2",
     "chai": "^4.3.7",
     "chai-json-schema": "^1.5.1",
     "dotenv": "^16.0.3",
     "eslint": "^8.32.0",
-    "eslint-config-standard-with-typescript": "^31.0.0",
-    "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-n": "^15.6.1",
-    "eslint-plugin-promise": "^6.1.1",
     "expect": "^29.3.1",
     "git-scripts": "^0.6.4",
     "mocha": "^10.2.0",
     "mochawesome": "^7.1.3",
     "mochawesome-report-generator": "^6.2.0",
     "prettier": "^2.8.3",
-    "randomstring": "^1.2.3",
     "schema-validator": "^3.3.1",
     "supertest": "^6.3.3",
     "ts-node": "^10.9.1",
